Use OffsetDateTime from azure_core::time (#472)
* Use OffsetDateTime from azure_core::time

Instead of the time crate.
Fixed an issue in the Use class that would cause a module to not be
brought into scope if types within that module are also brought into
scope.

* finish updating tests

* revert changes to Use class

it is now assumed that leaf values in calls to add() are a symbol within
a module and can never be a module.
updated calls to add() that were using a module to now use the required
leaf symbol.
consolidated cases for base64 and OffsetDateTime helpers.